+++ src/Phase 2/segmentation_models/models/fpn.py
@@ -160,7 +160,7 @@ def build_fpn(
         kernel_initializer='glorot_uniform',
         name='head_conv',
     )(x)
-    x = layers.Activation(activation, name=activation)(x)
+    x = layers.Activation(activation, dtype='float32', name=activation)(x)
 
     # create keras model instance
     model = models.Model(input_, x)
