+++ src/Phase 2/segmentation_models/models/unet.py
@@ -147,7 +147,7 @@ def build_unet(
         kernel_initializer='glorot_uniform',
         name='final_conv',
     )(x)
-    x = layers.Activation(activation, name=activation)(x)
+    x = layers.Activation(activation, dtype='float32',name=activation)(x)
     print(f"Sanity Check: Model Output = {x}")
     # create keras model instance
     model = models.Model(input_, x)
